LB6/requirements.txt
Flask==3.0.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
requests==2.31.0
